feature: Change LottoNumberGenerator util class to LottoMachine domain class with lotto number random generate responsibility
