+++ src/test/kotlin/lotto/LottoMachineTest.kt
@@ -4,12 +4,14 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.ints.shouldBeInRange
 import io.kotest.matchers.shouldBe
-import lotto.util.LottoNumberGenerator
+import lotto.domain.LottoMachine
 
-class LottoNumberGeneratorTest : StringSpec({
+class LottoMachineTest : StringSpec({
     "1~45 사이의 랜덤 숫자 6개를 반환한다."{
+        //given
+        val lottoMachine = LottoMachine()
         //when
-        val lottoNumbers = LottoNumberGenerator.generate(1)[0].toList()
+        val lottoNumbers = lottoMachine.generate(1)[0].toList()
         //then
         lottoNumbers.size shouldBe 6
         lottoNumbers.forAll { it shouldBeInRange 1..45 }
